% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{The fars_map_state function takes the state number and year as an input and displays a plot
of the state with the number of fatalaties mapped as dots at the appropriate location of the
incident. To be able to use this function ensure the maps package is installed}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{and years Take the statenum and year and displays a map plot of the state for the year specified with the dots for each fatal
injury incident}
}
\value{
This function returns map plot of the state for the year specified with the dots for
   each fatal injury incident with the map at their corresponding latitude and longitude of
   occurence


 @importFrom maps map
}
\description{
The fars_map_state function takes the state number and year as an input and displays a plot
of the state with the number of fatalaties mapped as dots at the appropriate location of the
incident. To be able to use this function ensure the maps package is installed
}
\examples{
fars_map_state(1,2013)
fars_map_state(4,2013)


}
