% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Checking the existance of a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{takes the file name as an input and checks if it exists}
}
\value{
This function either returns a message the filename does not exist if the file name


 @examples
 fars_read("accident_2013.csv.bz2")
 fars_read("accident_2017.csv.bz2")

 @importFrom readr read_csv
}
\description{
Checking the existance of a file
}
