% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{The fars_summarize_years function takes a year as an input and summarizes the number of
accidents by month within the input year specified}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{takes the years as input and displays a monthwise summary of the number of fatal injuries}
}
\value{
A table displaying a month number and the consolidated number of fatal injuries for
each month of the specified year
}
\description{
The fars_summarize_years function takes a year as an input and summarizes the number of
accidents by month within the input year specified
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2014)


}
