% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_make_filename.R
\name{make_filename}
\alias{make_filename}
\title{The make_filename function is used to create a filename}
\usage{
make_filename(year)
}
\description{
@param year Takes the year as an input and creates a filename
}
\details{
@return This function returns a character vector containing the formatted combination of text
    and variable values

@examples
 make_filename(2013)
 make_filename(2014)

@export
}
